+++ src/Modules/Profile/Profile.js
@@ -10,6 +10,10 @@ import { vueRegister, vueApply } from '@/vue'
 import BananaList from './components/BananaList'
 import PublicProfile from './components/PublicProfile'
 import PickupHistory from '../StoreUser/components/PickupHistory'
+// Wallpost
+import { URL_PART } from '@/browser'
+import '../WallPost/WallPost.css'
+import { initWall } from '@/wall'
 
 expose({ trySendBuddyRequest })
 
@@ -32,3 +36,8 @@ vueRegister({
 vueApply('#vue-profile-bananalist', true) // BananaList
 vueApply('#vue-pickup-history', true) // PickupHistory
 vueApply('#profile-public', true) // PublicProfile
+
+if (URL_PART(0) === 'profile') {
+  const wallpostTable = (URL_PART(2) === 'notes') ? 'usernotes' : 'foodsaver'
+  initWall(wallpostTable, URL_PART(1))
+}